Fix chat drawer and spacing in full page view. Also fix infect bug
There was a bug in fake server where members could be duplicate in the member list. Doesn't happen in firestore because we do an array union, but in fake server we have to actually check that the user isn't already in the array.

Fix for drawer was that it can't be nested in a div. That was rendering the drawer out of bounds and was a pain to debug.